+++ src/main/java/net/silentchaos512/berries/data/DataGenerators.java
@@ -3,7 +3,6 @@ package net.silentchaos512.berries.data;
 import net.minecraft.data.DataGenerator;
 import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.common.EventBusSubscriber;
-import net.neoforged.fml.common.Mod;
 import net.neoforged.neoforge.data.event.GatherDataEvent;
 
 @EventBusSubscriber(bus = EventBusSubscriber.Bus.MOD)
@@ -18,6 +17,12 @@ public final class DataGenerators {
         gen.addProvider(true, new ModItemModelProvider(event));
         gen.addProvider(true, new ModLanguageProvider(event.getGenerator()));
 
+        gen.addProvider(true, new ModLootTableProvider(event));
+        gen.addProvider(true, new ModGlobalLootModifierProvider(event));
+
+        var blockTagsUnused = new ModBlockTagsProvider(event);
+        gen.addProvider(true, blockTagsUnused);
+        gen.addProvider(true, new ModItemTagProvider(event, blockTagsUnused));
         gen.addProvider(true, new ModRecipeProvider(event));
 
         gen.addProvider(true, new WorldGenGenerator(event));