+++ src/main/java/net/silentchaos512/berries/data/ModBlockTagsProvider.java
@@ -1,0 +1,16 @@
+package net.silentchaos512.berries.data;
+
+import net.minecraft.core.HolderLookup;
+import net.neoforged.neoforge.common.data.BlockTagsProvider;
+import net.neoforged.neoforge.data.event.GatherDataEvent;
+import net.silentchaos512.berries.BerriesMod;
+
+class ModBlockTagsProvider extends BlockTagsProvider {
+    public ModBlockTagsProvider(GatherDataEvent event) {
+        super(event.getGenerator().getPackOutput(), event.getLookupProvider(), BerriesMod.MOD_ID, event.getExistingFileHelper());
+    }
+
+    @Override
+    protected void addTags(HolderLookup.Provider pProvider) {
+    }
+}